Delete sync rate mechanism
