+++ app/protocol/flows/blockrelay/handle_relay_invs.go
@@ -33,7 +33,6 @@ type RelayInvsContext interface {
 	IsIBDRunning() bool
 	TrySetIBDRunning() bool
 	UnsetIBDRunning()
-	UpdateRecentBlockAddedTimesWithLastBlock()
 }
 
 type handleRelayInvsFlow struct {
